Adding support for Proc objects.
If a Proc object (or a lambda) is set as the value of a configuration
variable, it will be "called" when that value is fetched. This is great for
delayed evaluation of configuration settings. For instance, when one
configuration setting is a composite of a few other ones (URL is composed of
the host/port/path) -- stuff like that.