Adding a default "undefined" value for configuration properties.
If a user attempts to use a configuration property that has not been set, they
will often times end up with a MethodMissing error. This occurs when accessing
a property that has not been set; the value returned is nil in this case.
Attempting to access sub-properties of nil is no good.