The Iterator class now skips Undefined objects in the Configuration.
Undefined objects end up in the Configuration when an attempt is make to
access a value that does not exist. These are used to print error messages to
the screen saying "hey, that does not exist". Unfortuantely the Undefined
objects were also showing up in help messages via the iterator.

This patch causes the iterator to ignore these Undefined objects when it finds
them.