Fixing specs so they actually test our functionality
Because Loquacious::Undefined objects have no methods defined on them, all
the RSpec "should" methods get consumed when applied to udnefiend return
values. Any spec assert will be true because the assertion never happens.

We change this style ...

  cfg.one.two.should be == 'foo'

into this style ...

  cfg.one.two.eql?('foo').should be_true

and all is better. This caught a few bugs in the loquacious system with
respect to merging configurations.